改使用summer-ci.js文件

+++ src/main/webapp/static/js/FrmWelcome.js
@@ -1,33 +1,29 @@
-import * as ci from "./SummerCI.js";
-
-export default class FrmWelcome extends ci.TPage {
-    ds = new ci.DataSet();
-    grid = new ci.TGrid().setId('grid').setDataSet(this.ds).setBorder('1px');
-    edtCode = new ci.TInput().setId('edtCode').setName('edtCode');
-    btnDelete = new ci.TButton().setId('btnDelete').setText("删除");
-    btnShow = new ci.TButton().setId('btnShow').setText("远程");
-    btnAppend = new ci.TButton().setId('btnAppend').setText("增加");
+export default class FrmWelcome extends sci.TPage {
+    ds = new sci.DataSet();
+    grid = new sci.TGrid().setId('grid').setDataSet(this.ds).setBorder('1px');
+    edtCode = new sci.TInput().setId('edtCode').setName('edtCode');
+    btnDelete = new sci.TButton().setId('btnDelete').setText("删除");
+    btnShow = new sci.TButton().setId('btnShow').setText("远程");
+    btnAppend = new sci.TButton().setId('btnAppend').setText("增加");
 
     constructor(owner) {
         super(owner);
 
         this.btnDelete.setOwner(this);
-        new ci.TSpan(this).setText(' ');
+        new sci.TSpan(this).setText(' ');
         this.btnShow.setOwner(this);
-        new ci.TLine(this);
+        new sci.TLine(this);
 
         this.grid.setOwner(this);
 
-        let form = new ci.TDiv(this);
+        let form = new sci.TDiv(this);
         this.edtCode.setOwner(form).writerProperty('value', 'hello');
-        new ci.TSpan(form).setText(' ');
+        new sci.TSpan(form).setText(' ');
         this.btnAppend.setOwner(form);
-    }
 
-    registerEvents() {
-        this.addEventListener('btnShow', 'click', this.btnShowClick);
-        this.addEventListener('btnAppend', 'click', this.btnAppendClick);
-        this.addEventListener('btnDelete', 'click', this.btnDeleteClick);
+        this.btnShow.addEventListener('click', this.btnShowClick);
+        this.btnAppend.addEventListener('click', this.btnAppendClick);
+        this.btnDelete.addEventListener('click', this.btnDeleteClick);
     }
 
     getDataSet() {
@@ -35,27 +31,27 @@ export default class FrmWelcome extends ci.TPage {
     }
 
     btnShowClick = () => {
-        let svr = new ci.RemoteService(this);
+        let svr = new sci.RemoteService(this);
         svr.setService('SvrExample.search');
-        svr.exec(() => {
-            let json = svr.getDataOut().getJson();
-            this.ds.setJson(json);
-            this.ds = new ci.DataSet(json);
-            this.grid.setDataSet(this.ds);
+        svr.exec((dataOut) => {
+            this.ds = dataOut;
+            this.grid.setDataSet(dataOut).addColumns(dataOut.getFieldDefs());
             this.grid.render();
         });
     };
 
     btnAppendClick = () => {
         this.ds.append();
-        this.ds.setField('corpNo_', 'd');
-        this.ds.setField('name_', this.edtCode.getHtmlValue());
-        this.ds.setField('size', this.ds.size());
+        this.ds.setValue('corpNo_', 'd');
+        this.ds.setValue('name_', this.edtCode.getHtmlValue());
+        this.ds.setValue('size', this.ds.size());
         this.grid.render();
     };
 
     btnDeleteClick = () => {
-        this.ds.delete();
+        this.ds.last();
+        if (!this.ds.eof())
+            this.ds.delete();
         this.grid.render();
 
     };
